Day 17-1
You know, I was in the middle of creating a heap when my original solution (node-indexed array) finished, let's go! It only took like 10000 times longer than the example input.

After some more time working on it, it works! It's still about as slow as before, but it does work, so yay?